avito_parser/requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.7
SQLAlchemy==2.0.41
telebot==0.0.5
typing_extensions==4.14.0
urllib3==2.4.0
